% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_data_functions.r
\name{ceip_add_population}
\alias{ceip_add_population}
\title{Add eurostat population numbers to CEIP data frame. Warning: only available from 2007!}
\usage{
ceip_add_population(df)
}
\arguments{
\item{df}{CEIP data frame generated by ceip_read() (of class ceipr_data)}
}
\value{
Merges the two databases
}
\description{
Add eurostat population numbers to CEIP data frame. Warning: only available from 2007!
}
\keyword{emission,}
\keyword{population}
\keyword{voc,}
