% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceip_animation.r
\name{ceip_animation}
\alias{ceip_animation}
\title{Animate a raster stack of Centre on Emission Inventories
and Projections (CEIP) data into an animated gif for terrestrial locations.}
\usage{
ceip_animation(file = "~/SOx_A.tif", output_file = "~/test.gif",
  region = "belgium", width = 1200, height = 1200, xlim = c(-30,
  90), ylim = c(30, 82), upper_limit = 15000, time_step = 0.5)
}
\arguments{
\item{file}{CEIP txt data file}

\item{output_file}{path to a gif file to store the final animation}

\item{region}{which regions to consider (default = "Belgium")}

\item{width}{output width in px (default = 800)}

\item{height}{output height in px (default = 800)}

\item{xlim}{longitudinal geographic limits, override the regions argument
(default = c())}

\item{ylim}{latitudinal geographic limits, override the regions argument
(default = c())}

\item{upper_limit}{maximum value of scale (default = 10)}

\item{time_step}{pause in seconds between two gif frames}
}
\value{
geotiff written to disk or raster data returned to R workspace
when written to disk the location of the txt file is used
}
\description{
Animate a raster stack of Centre on Emission Inventories
and Projections (CEIP) data into an animated gif for terrestrial locations.
}
\keyword{animation}
\keyword{emission,}
\keyword{plot,}
\keyword{voc,}
