% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.r
\name{ceip_comparisons}
\alias{ceip_comparisons}
\title{Compare CEIP emissions between years}
\usage{
ceip_comparisons(df, group = "iso2", year = 2000,
  group_label = "country", subject = "", title = paste(subject,
  "comparison per", group_label, sep = " "))
}
\arguments{
\item{df}{CEIP data frame generated by ceip_read() (of class ceipr_data)}

\item{group}{group to sort by (default = "iso2")}

\item{year}{year to evaluate (default = 2000)}

\item{group_label}{group label to use (default = "country")}

\item{subject}{subject line (default = "")}

\item{title}{plot title}
}
\value{
ggplot object to be plotted
}
\description{
Compare CEIP emissions between years
}
\keyword{emission,}
\keyword{plot,}
\keyword{trend}
\keyword{voc,}
