% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{ceip_data_file}
\alias{ceip_data_file}
\title{Formats the name of the CEIP file inside the zip file}
\usage{
ceip_data_file(year, pollutant, sector)
}
\arguments{
\item{year}{year to download}

\item{pollutant}{different pollutants}

\item{sector}{CEIP sectors}
}
\value{
url location
}
\description{
Formats the name of the CEIP file inside the zip file
}
