% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceip_download.r
\name{ceip_download}
\alias{ceip_download}
\title{Download zipped Centre on Emission Inventories and Projections (CEIP) data}
\usage{
ceip_download(pollutant = c("SOx", "NOx"), year = 2000:2016,
  out_dir = tempdir(),
  base_url = "http://webdab1.umweltbundesamt.at/download/gridding2018/",
  silent = FALSE)
}
\arguments{
\item{out_dir}{output directory (default = tempdir())}

\item{base_url}{source url to download data from}

\item{pollutants}{character string with pollutants
(CO, NH3, NMVOC, NOx, PM2_5, PM10, PMcoarse, SOx)}

\item{years}{years to download (all = 2000:2016)}
}
\value{
Downloads ceip zip files into the output directory. CRAN policy does
not allow default settings to write to the userspace so tempdir() is the
default. However, a different directory can be used for caching.
}
\description{
The download routine will recreate the CEIP file structure in the output
directory (sorting files in folders by year).
}
\examples{

\dontrun{
ceip_download(pollutant = "SOx",
              year = 2000)
}
}
\keyword{emission,}
\keyword{voc}
