% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_data_functions.r
\name{ceip_population_emissions}
\alias{ceip_population_emissions}
\title{Calculate per capita emissions}
\usage{
ceip_population_emissions(df)
}
\arguments{
\item{df}{CEIP data frame generated by ceip_read() (of class ceipr_data)
and supplemented with eurostat data using ceip_add_population()}
}
\value{
Merges the two databases
}
\description{
Calculate per capita emissions
}
\keyword{emission,}
\keyword{population}
\keyword{voc,}
