% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceip_regression.r
\name{ceip_regression}
\alias{ceip_regression}
\title{Run a simple linear regression on the
the Centre on Emission Inventories and Projections (CEIP)
emission data as formatted using ceip_download()}
\usage{
ceip_regression(file = "~/SOx_A.tif", out_dir = "~", plot = FALSE)
}
\arguments{
\item{file}{path to a file generated by ceip_download()}

\item{out_dir}{output directory}

\item{internal}{logical TRUE / FALSE}
}
\value{
geotiff written to disk
}
\description{
Run a simple linear regression on the
the Centre on Emission Inventories and Projections (CEIP)
emission data as formatted using ceip_download()
}
\examples{

\dontrun{
ceip_download()
}
}
\keyword{emission,}
\keyword{voc}
