% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{ceip_sector_meta_data}
\alias{ceip_sector_meta_data}
\title{Sector meta data (code, name, etc)}
\usage{
ceip_sector_meta_data()
}
\value{
sector meta data
}
\description{
Sector meta data (code, name, etc)
}
