% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceip_to_raster.r
\name{ceip_to_raster}
\alias{ceip_to_raster}
\title{Converts text based data from the
Centre on Emission Inventories and Projections (CEIP)
into raster grid data (geotiff or raster object in the R workspace)}
\usage{
ceip_to_raster(df, year = 2000:2016, pollutant = "SOx", sector = "A",
  country = NULL, out_dir = tempdir(), trim = TRUE,
  internal = TRUE)
}
\arguments{
\item{df}{CEIP data frame generated by ceip_read()}

\item{year}{which years to cover (default = 2000:2016)}

\item{pollutant}{emission values to include (default = "NOx")}

\item{sector}{sector to visualize}

\item{country}{country to visualize, if NULL all (default = \code{NULL})}

\item{out_dir}{directory where to store the geotiff if written to file
(default = tempdir())}

\item{trim}{shrink map to the area covered by data (default = \code{TRUE})}

\item{internal}{bolean \code{TRUE} / \code{FALSE}, write output to disk or
return raster stack to R workspace}
}
\value{
Returns a geotiff written to disk or raster stack returned
to the R workspace when written to disk the out_dir location is used.
}
\description{
Converts text based data from the
Centre on Emission Inventories and Projections (CEIP)
into raster grid data (geotiff or raster object in the R workspace)
}
\examples{

\dontrun{
ceip_grid()
}
}
\keyword{emission,}
\keyword{voc}
