% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_data_functions.r
\name{ceip_totals}
\alias{ceip_totals}
\title{Summarize emission totals}
\usage{
ceip_totals(df)
}
\arguments{
\item{df}{CEIP data frame generated by ceip_read() (of class ceipr_data)}
}
\value{
a dataframe with emissions per country
}
\description{
Summarize emission totals
}
