% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_functions.r
\name{ceip_trends}
\alias{ceip_trends}
\title{Visualize CEIP trends in emissions}
\usage{
ceip_trends(df, group = "iso2", group_label = "country",
  subject = "", title = paste(subject, "trends per", group_label, sep =
  " "))
}
\arguments{
\item{df}{CEIP data frame generated by ceip_read() (of class ceipr_data)}

\item{group}{group to sort by (default = "iso2")}

\item{group_label}{group label to use (default = "country")}

\item{subject}{subject line (default = "")}

\item{title}{plot title}
}
\value{
ggplot object to be plotted
}
\description{
Visualize CEIP trends in emissions
}
\keyword{emission,}
\keyword{plot,}
\keyword{trend}
\keyword{voc,}
