% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{ceip_zip_file}
\alias{ceip_zip_file}
\title{Formats the name of the CEIP zip file location}
\usage{
ceip_zip_file(path, year, pollutant)
}
\arguments{
\item{path}{CEIP sectors}

\item{year}{year to download}

\item{pollutant}{different pollutants}
}
\value{
zip file location string
}
\description{
location is either remote or local as defined
by the path variable
}
