% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_data_functions.r
\name{check_for_totals}
\alias{check_for_totals}
\title{Add eurostat population numbers check}
\usage{
check_for_totals(df)
}
\arguments{
\item{df}{CEIP data frame generated by ceip_totals()}
}
\value{
warning / stop message upon failure
}
\description{
Add eurostat population numbers check
}
\keyword{checks}
\keyword{data}
\keyword{emission,}
\keyword{population,}
\keyword{voc,}
