% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{convert_to_raster}
\alias{convert_to_raster}
\title{converts ceip subset data frame to a raster}
\usage{
convert_to_raster(df)
}
\arguments{
\item{df}{subset of a full ceip data frame as returned by ceip_read()}
}
\value{
raster layer
}
\description{
Never to be used stand alone, only as a helper function
}
