% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{maps_to_sp}
\alias{maps_to_sp}
\title{Helper function to slice and dice a map and convert it to an sp() object}
\usage{
maps_to_sp(regions = "Belgium", xlim = c(2.4, 6.5), ylim = c(49, 52),
  length_out = 100, clip = TRUE)
}
\arguments{
\item{regions}{list of regions (countries to convert)}

\item{xlim}{longitude bounding box values}

\item{ylim}{latitude bounding box values}

\item{length_out}{segment output length}

\item{clip}{bolean TRUE / FALSE, clip polygons on bounding box}
}
\value{
an spatialPolygon object which can be used for subsetting, and
the bounding box constraining the object
}
\description{
Helper function to slice and dice a map and convert it to an sp() object
}
\keyword{emission,}
\keyword{polygon,}
\keyword{subsetting}
\keyword{voc,}
