% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceip_download.r
\name{ceip_download}
\alias{ceip_download}
\title{Download and compile emission data from
the Centre on Emission Inventories and Projections (CEIP)}
\usage{
ceip_download(pollutants = c("SOx"), sectors = "A", out_dir = "~",
  base_url = "http://webdab1.umweltbundesamt.at/download/gridding2018/")
}
\arguments{
\item{pollutants}{character string with pollutants
(SOx, NOx, PM2_5, PM10, )}

\item{sectors}{different societal sectors for which to compile the data
as referenced by their abbreviation (A - N & NT)}

\item{out_dir}{output directory}

\item{base_url}{source url to download data from}
}
\value{
geotiff written to disk
}
\description{
Download and compile emission data from
the Centre on Emission Inventories and Projections (CEIP)
}
\examples{

\dontrun{
ceip_download()
}
}
\keyword{emission,}
\keyword{voc}
