% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceip_grid.r
\name{ceip_grid}
\alias{ceip_grid}
\title{Converts text based data from the 
Centre on Emission Inventories and Projections (CEIP)
into raster data (geotiff or raster object in the R workspace)}
\usage{
ceip_grid(file, out_dir = tempdir(), internal = FALSE)
}
\arguments{
\item{file}{CEIP txt data file}

\item{out_dir}{output directory where to store data when writing to disk}

\item{internal}{bolean TRUE / FALSE, write output to disk or return to
R workspace}
}
\value{
geotiff written to disk or raster data returned to R workspace
}
\description{
Converts text based data from the 
Centre on Emission Inventories and Projections (CEIP)
into raster data (geotiff or raster object in the R workspace)
}
\examples{

\dontrun{
ceip_grid()
}
}
\keyword{emission,}
\keyword{voc}
