% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceip_read.r
\name{ceip_read}
\alias{ceip_read}
\title{Load CEIP zipped data into a tidy data frame}
\usage{
ceip_read(pollutant = "NOx", sector = c("A", "B"), year = 2000:2016,
  path = "~/Desktop/tmp/")
}
\arguments{
\item{pollutant}{different pollutants}

\item{sector}{CEIP sectors}

\item{year}{which years to include in the data compilation}

\item{path}{path with original CEIP zip files}
}
\value{
Returns a data frame (tibble) of CEIP data. This data is tidy
and can be easily used in statistical analysis. Or converted to geospatial
data using included functions.
}
\description{
Load data for specified pollutant, for a specific sector and all years from
a directory of zip files, or an online path. For the latter option your,
milage might vary and it is recommended to download files of interest first
using the ceip_download() function as it is less likely to cause time-out
issues and will limit overall traffic to the server.
}
