% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{ceip_read_zip}
\alias{ceip_read_zip}
\title{Reads internal file from a zip file (unzip in memory)}
\usage{
ceip_read_zip(zip, filename)
}
\arguments{
\item{zip}{location of a zip file (url / or local path)}

\item{filename}{filename in the zip file to extract}
}
\value{
reads a specific file
}
\description{
Reads internal file from a zip file (unzip in memory)
}
